feat: add content-type header for direct attachments
